fetch workspaces icp + set workspaces
